feat: completed the README and fine-tuned the project structure
