--- src/.env
@@ -1,18 +1,0 @@
-# HOST
-VITE_HOST_API=http://43.136.232.116/
-
-# ASSETS
-# VITE_ASSETS_API=https://api-dev-minimal-v510.vercel.app
-
-# GAO DE MAP
-VITE_GAODE_ADMIN_API=https://restapi.amap.com/v3/config/district
-VITE_GAODE_IP_API=https://restapi.amap.com/v3/ip?key=
-VITE_GAODE_API_KEY=03eceb9420e057a98616285039c15367
-
-# FIREBASE
-VITE_FIREBASE_API_KEY=
-VITE_FIREBASE_AUTH_DOMAIN=
-VITE_FIREBASE_PROJECT_ID=
-VITE_FIREBASE_STORAGE_BUCKET=
-VITE_FIREBASE_MESSAGING_SENDER_ID=
-VITE_FIREBASE_APPID=
