+++ src/apis/index.js
@@ -1,0 +1,23 @@
+// API
+// ----------------------------------------------------------------------
+
+// Base URL
+export const HOST_API = 'http://43.136.232.116/'
+
+// ----------------------------------------------------------------------
+
+export const GAODE_API = {
+    apiKey: '03eceb9420e057a98616285039c15367',
+    apiAdmin: 'https://restapi.amap.com/v3/config/district',
+    apiIP: 'https://restapi.amap.com/v3/ip?key='
+};
+
+// ----------------------------------------------------------------------
+
+// request for the backend
+export const endpoints = {
+    auth: {
+        login: '/api/user/login',
+        register: '/api/user/register',
+    },
+};
