--- src/assets/data/index.js
@@ -1,1 +1,0 @@
-export * from './countries';
