+++ src/components/city-cascader/utils.js
@@ -1,5 +1,6 @@
 import axios from 'axios';
-import provinceCityMapping from '@/utils/zh-en'
+import provinceCityMapping from 'src/utils/zh-en';
+import { GAODE_API } from 'src/apis/index';
 
 // translate function
 export function translateName(chineseName) {
@@ -15,9 +16,9 @@ export const reverseMapping = Object.keys(provinceCityMapping).reduce((acc, chin
 
 // get province by Gao De api
 export async function fetchProvinces() {
-    const response = await axios.get(`https://restapi.amap.com/v3/config/district`, {
+    const response = await axios.get(GAODE_API.apiAdmin, {
         params: {
-            key: '03eceb9420e057a98616285039c15367',
+            key: GAODE_API.apiKey,
             keywords: '中国',
             subdistrict: 1,
         },
@@ -34,9 +35,9 @@ export async function fetchCitiesByProvince(provinceName) {
     // Use the reverse mapping to find the Chinese name corresponding to the English name
     const chineseProvinceName = reverseMapping[provinceName] || provinceName;
 
-    const response = await axios.get(`https://restapi.amap.com/v3/config/district`, {
+    const response = await axios.get(GAODE_API.apiAdmin, {
         params: {
-            key: '03eceb9420e057a98616285039c15367',
+            key: GAODE_API.apiKey,
             keywords: chineseProvinceName,
             subdistrict: 1,
         },
