--- src/config-global.js
@@ -1,25 +1,0 @@
-// API
-// ----------------------------------------------------------------------
-
-// base URL
-export const HOST_API = 'http://43.136.232.116/'
-
-// Gao De API
-export const GAODE_API = {
-    apiKey: import.meta.env.VITE_GAODE_API_KEY,
-    apiAdmin: import.meta.env.VITE_GAODE_ADMIN_API,
-    apiIP: import.meta.env.VITE_GAODE_IP_API
-};
-
-// ROOT PATH AFTER LOGIN SUCCESSFUL
-export const PATH_AFTER_LOGIN = '/';
-
-// export const FIREBASE_API = {
-//   apiKey: import.meta.env.VITE_FIREBASE_API_KEY,
-//   authDomain: import.meta.env.VITE_FIREBASE_AUTH_DOMAIN,
-//   projectId: import.meta.env.VITE_FIREBASE_PROJECT_ID,
-//   storageBucket: import.meta.env.VITE_FIREBASE_STORAGE_BUCKET,
-//   messagingSenderId: import.meta.env.VITE_FIREBASE_MESSAGING_SENDER_ID,
-//   appId: import.meta.env.VITE_FIREBASE_APPID,
-//   measurementId: import.meta.env.VITE_FIREBASE_MEASUREMENT_ID,
-// };
