--- src/pages/Login/index.scss
@@ -1,191 +1,0 @@
-.login {
-  width: 100%;
-  height: 100%;
-  position: absolute;
-  left: 0;
-  top: 0;
-
-  display: flex;
-
-  .flex-column {
-    display: flex;
-    flex-direction: column;
-  }
-
-  .flex-row {
-    display: flex;
-    flex-direction: row;
-  }
-
-  .login-left {
-    width: 54%;
-    background: center/cover url('@/assets/login.png');
-  }
-
-  .login-right {
-    width: 46%;
-    background-color: #FFFFFF;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-  }
-
-  @media (max-width: 900px) {
-    .login-left {
-      display: none;
-    }
-
-    .login-right {
-      width: 100%;
-    }
-  }
-
-  .login-right-title {
-    width: 75%;
-    height: 55px;
-    display: flex;
-    justify-content: left;
-    align-items: center;
-  }
-
-  .login-right-title p {
-    height: 100%;
-    display: flex;
-    align-items: center;
-    color: rgb(33, 43, 54);
-    font-size: 24px;
-    font-weight: 700;
-  }
-
-  .login-right-signup {
-    width: 75%;
-    height: 36px;
-    align-items: center;
-  }
-
-  .login-right-signup p {
-    margin-right: 5px;
-  }
-
-  .login-right-signup a {
-    color: rgb(0, 167, 111);
-    text-decoration: none;
-  }
-
-  .login-right-signup a:hover {
-    text-decoration: underline;
-  }
-
-  .login-right-input {
-    width: 75%;
-    margin-top: 15px;
-  }
-
-  .login-right-input-div {
-    margin-top: 20px;
-    width: 100%;
-    height: 50px;
-    display: flex;
-  }
-
-  .input-animate {
-    position: relative;
-    color: rgb(147, 158, 170);
-  }
-
-  .input-animate input[type='text'],
-  .input-animate input[type='password'] {
-    height: 50px;
-    border: 1px solid rgb(233, 236, 238);
-    border-radius: 8px;
-    padding-left: 10px;
-    font-size: 17px;
-    transition: border-color 0.1s ease-in-out;
-  }
-
-  .input-animate input[type='text']:hover,
-  .input-animate input[type='password']:hover {
-    outline: none;
-    border: 1px solid rgb(33, 43, 54);
-  }
-
-  .input-animate input[type='text']:focus,
-  .input-animate input[type='password']:focus {
-    outline: none;
-    border: 1px solid rgb(33, 43, 54);
-  }
-
-  .input-animate label {
-    padding-left: 5px;
-    padding-right: 5px;
-    display: flex;
-    align-items: center;
-    position: absolute;
-    left: 11.25px;
-    top: 15.5px;
-    font-size: 12px;
-    pointer-events: none;
-  }
-
-  .input-animate input[type='text']:focus+label,
-  .input-animate input[type='password']:focus+label {
-    top: -7px;
-    font-size: 13px;
-    color: rgb(33, 43, 54);
-    transition: 0.3s;
-    justify-content: center;
-    background-color: white;
-  }
-
-  .login-right-forget {
-    margin-top: 10px;
-    width: 75%;
-    height: 50px;
-    display: flex;
-    justify-content: right;
-    align-items: center;
-  }
-
-  .login-right-forget a {
-    color: rgb(33, 43, 54);
-  }
-
-  .login-right-button {
-    width: 75%;
-    margin-top: 10px;
-  }
-
-  .login-right-button button {
-    width: 100%;
-    height: 48px;
-    background-color: rgb(33, 43, 54);
-    border-radius: 8px;
-    color: #FFFFFF;
-    font-weight: 700;
-    font-size: 15px;
-  }
-
-  .login-logo {
-    width: 200px;
-    height: 60px;
-    display: block;
-    margin: 0 auto 20px;
-  }
-
-  .login-container {
-    width: 440px;
-    height: 360px;
-    position: absolute;
-    left: 50%;
-    top: 50%;
-    transform: translate(-50%, -50%);
-    box-shadow: 0 0 50px rgb(0 0 0 / 10%);
-  }
-
-  .login-checkbox-label {
-    color: #1890ff;
-  }
-
-
-}
