+++ src/utils/axios.js
@@ -1,38 +1,40 @@
 import axios from 'axios';
-import { HOST_API } from 'src/config-global';
+import { HOST_API } from 'src/apis/index';
 
 // ----------------------------------------------------------------------
+
+// Can be used to send various requests get post delete
 const axiosInstance = axios.create({ baseURL: HOST_API });
 
 axiosInstance.interceptors.response.use(
   (res) => res,
   (error) => {
-    // 如果有响应并且响应中包含数据
+    // If there is a response and the response contains data
     if (error.response && error.response.data) {
-      // 构造一个自定义错误对象
+      // Constructing a custom error object
       let customError = {
         status_code: error.response.data.status_code,
         status_msg: error.response.data.status_msg,
         data: {}
       };
 
-      // 如果有剩余尝试次数，添加到错误对象中
+      // If there is a remaining number of attempts, add to the error object
       if (error.response.data.Data.remaining_attempts !== undefined) {
         customError.data.remainingAttempts = error.response.data.Data.remaining_attempts;
       }
 
-      // 如果账号被锁定，添加锁定到期时间
+      // If the account is locked, add the lock expiration time
       if (error.response.data.Data.lock_expires !== undefined) {
-        // 将时间戳转换为更易读的格式，如果需要
+        // Convert timestamps to a more readable format, if desired
         const expires = new Date(parseInt(error.response.data.Data.lock_expires) * 1000);
         customError.data.lockExpires = expires.toString();
       }
 
-      // 返回自定义错误对象，而不是抛出异常
+      // Returning custom error objects instead of throwing exceptions
       return Promise.reject(customError);
     }
     
-    // 如果没有响应或者响应中不包含数据，返回通用错误信息
+    // If there is no response or the response does not contain data, return a generic error message
     return Promise.reject((error.response && error.response.data)||'Something went wrong');
   }
 );
@@ -40,7 +42,8 @@ axiosInstance.interceptors.response.use(
 export default axiosInstance;
 
 // ----------------------------------------------------------------------
-// 获取get请求
+
+// Use only for get requests, Simplifies code, Increases maintainability
 
 export const fetcher = async (args) => {
   const [url, config] = Array.isArray(args) ? args : [args];
@@ -48,33 +51,4 @@ export const fetcher = async (args) => {
   const res = await axiosInstance.get(url, { ...config });
 
   return res.data;
-};
-
-// ----------------------------------------------------------------------
-
-export const endpoints = {
-  //   chat: '/api/chat',
-  //   kanban: '/api/kanban',
-  //   calendar: '/api/calendar',
-  auth: {
-    me: '/api/auth/me',
-    login: '/api/user/login',
-    register: '/api/user/register',
-  },
-  //   mail: {
-  //     list: '/api/mail/list',
-  //     details: '/api/mail/details',
-  //     labels: '/api/mail/labels',
-  //   },
-  //   post: {
-  //     list: '/api/post/list',
-  //     details: '/api/post/details',
-  //     latest: '/api/post/latest',
-  //     search: '/api/post/search',
-  //   },
-  //   product: {
-  //     list: '/api/product/list',
-  //     details: '/api/product/details',
-  //     search: '/api/product/search',
-  //   },
 };
