--- src/utils/index.js
@@ -1,12 +1,0 @@
-// 统一中转工具函数模块
-// import {request} from @/utils
-
-import { request } from './request'
-import { setToken, getToken, removeToken } from './token'
-
-export {
-  request,
-  setToken,
-  getToken,
-  removeToken
-}
